Moving messages to their own module
