+++ web/messages/__init__.py
@@ -1,0 +1,1 @@
+"""Contain the messages for the web app."""
